refactor: remove old create app modal file
