--- src/components/app/create/createApp.scss
@@ -1,55 +1,0 @@
-/*
- * Copyright (c) 2024. Devtron Inc.
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-.create-app-container {
-    .scrollable-content {
-        height: calc(100vh - 120px);
-        overflow: auto;
-
-        .info-container--create-app {
-            padding: 10px 12px;
-            align-items: start;
-            grid-template-columns: 20px 1fr auto;
-            grid-column-gap: 12px;
-
-            .dc__info-title {
-                display: inline;
-            }
-        }
-
-        .form__row {
-            margin-bottom: 16px;
-            span.form__label {
-                font-size: 12px;
-                color: var(--N700);
-            }
-
-            .form__text-field-info--create-app {
-                align-items: flex-start;
-
-                .form__icon--create-app {
-                    position: relative;
-                    top: 3px;
-                }
-            }
-        }
-    }
-
-    .mandatory-tag-key-container {
-        min-width: 345px;
-        width: 345px;
-    }
-}
