+++ src/components/app/details/AboutTagEditModal.tsx
@@ -37,7 +37,6 @@ import { ReactComponent as Close } from '../../../assets/icons/ic-cross.svg'
 import { AboutAppInfoModalProps } from '../types'
 import { editApp } from '../service'
 import { importComponentFromFELibrary } from '../../common'
-import '../create/createApp.scss'
 import { APP_TYPE } from '@Config/constants'
 import { getLabelTags } from './utils'
 
@@ -182,7 +181,7 @@ export default function AboutTagEditModal({
 
     return (
         <Drawer position="right" width="800px" onClose={onClose} onEscape={onClose}>
-            <div className="h-100 bg__primary create-app-container" onClick={stopPropagation}>
+            <div className="h-100 bg__primary" onClick={stopPropagation}>
                 <div className="flex dc__content-space pt-16 pb-16 pl-20 pr-20 dc__border-bottom">
                     <h2 className="fs-16 cn-9 fw-6 m-0">Manage tags</h2>
                     <Close className="icon-dim-20 cursor" onClick={onClose} />
