+++ src/components/v2/values/chartValuesDiff/ChartValuesView.tsx
@@ -111,7 +111,6 @@ import {
     ChartValuesViewType,
 } from './ChartValuesView.type'
 import { chartValuesReducer, initState } from './ChartValuesView.reducer'
-import { ValidationRules } from '../../../app/create/validationRules'
 import { getAndUpdateSchemaValue, updateGeneratedManifest } from './ChartValuesView.utils'
 import { getAppId } from '../../appDetails/k8Resource/nodeDetail/nodeDetail.api'
 import ChartValuesGUIForm from './ChartValuesGUIView'
@@ -139,6 +138,7 @@ import ClusterNotReachableDailog from '../../../common/ClusterNotReachableDailog
 import { VIEW_MODE } from '@Pages/Shared/ConfigMapSecret/constants'
 import IndexStore from '../../appDetails/index.store'
 import { AUTO_GENERATE_GITOPS_REPO, CHART_VALUE_ID } from './constant'
+import { validateAppName } from '@Components/app/CreateAppModal/utils'
 
 const GeneratedHelmDownload = importComponentFromFELibrary('GeneratedHelmDownload')
 const getDownloadManifestUrl = importComponentFromFELibrary('getDownloadManifestUrl', null, 'function')
@@ -202,7 +202,6 @@ const ChartValuesView = ({
 
     const [obj] = useJsonYaml(commonState.modifiedValuesYaml, 4, 'yaml', true)
     const isUpdate = isExternalApp || (commonState.installedConfig?.environmentId && commonState.installedConfig.teamId)
-    const validationRules = new ValidationRules()
     const [showUpdateAppModal, setShowUpdateAppModal] = useState(false)
 
     const isPresetValueView =
@@ -805,7 +804,7 @@ const ChartValuesView = ({
     }
 
     const isValidData = (validatedAppName?: { isValid: boolean; message: string }) => {
-        const _validatedAppName = validatedAppName || validationRules.appName(appName)
+        const _validatedAppName = validatedAppName || validateAppName(appName)
 
         if (
             isDeployChartView &&
@@ -910,7 +909,7 @@ const ChartValuesView = ({
             return
         }
 
-        const validatedName = validationRules.appName(isCreateValueView ? valueName : appName)
+        const validatedName = validateAppName(isCreateValueView ? valueName : appName)
         if (!isRequestDataValid(validatedName)) {
             // If some validation error occurred, close the readme column or comparision column
             // to show the validations errors
@@ -1124,7 +1123,7 @@ const ChartValuesView = ({
     const handleTabSwitch = (e) => {
         if (e?.target && e.target.value !== commonState.activeTab) {
             if (e.target.value === 'manifest') {
-                const validatedName = validationRules.appName(isCreateValueView ? valueName : appName)
+                const validatedName = validateAppName(isCreateValueView ? valueName : appName)
                 if (isCreateValueView && !validatedName.isValid) {
                     dispatch({
                         type: ChartValuesViewActionTypes.multipleOptions,
@@ -1428,7 +1427,7 @@ const ChartValuesView = ({
     }
 
     const handleAppNameChange = (newAppName: string) => {
-        const validatedAppName = validationRules.appName(newAppName)
+        const validatedAppName = validateAppName(newAppName)
         if (!validatedAppName.isValid && commonState.invalidAppNameMessage !== validatedAppName.message) {
             dispatch({
                 type: ChartValuesViewActionTypes.multipleOptions,
@@ -1532,7 +1531,7 @@ const ChartValuesView = ({
         )
     }
     const handleValueNameChange = (newValueName: string) => {
-        const validatedValueName = validationRules.appName(newValueName)
+        const validatedValueName = validateAppName(newValueName)
         if (!validatedValueName.isValid && commonState.invalidAppNameMessage !== validatedValueName.message) {
             dispatch({
                 type: ChartValuesViewActionTypes.multipleOptions,
